requirements.txt
flask
python-dotenv
supabase==2.10.0
google-generativeai
bcrypt
